move car util functions to car_util.rs
