+++ src/utils/db/mod.rs
@@ -3,6 +3,7 @@
 
 pub mod car_index;
 pub mod car_stream;
+pub mod car_util;
 pub mod file_backed_obj;
 
 use async_trait::async_trait;
